Consolidate cross-crate test scaffolding into deslop-test-support
Introduce a dev-only deslop-test-support crate as the shared home for E2E
test helpers the three binary crates' suites each duplicated: the version
manifest assertion, the LSP Content-Length frame read/write, the deslop-lsp
spawn idiom, and the with_ext report-path helper. Each local helper now
delegates to the shared crate (name/signature unchanged, no call-site churn).

Also fold the three deslop tests with a local 2-arg run_report onto the
existing common run_report(scan_root, min_nodes), and extract the repeated
spawn+handshake+initial-report preamble in notifications.rs into one helper.

Pure DRY refactor of test scaffolding; no behaviour change. The examples/
clone-demonstration corpus is left duplicated on purpose (documented in
examples/README.md).
