+++ crates/deslop-mcp/tests/wrong_root.rs
@@ -11,7 +11,7 @@
 
 use std::{
     fs,
-    io::{BufRead, BufReader, Read, Write},
+    io::{BufRead, BufReader, Write},
     path::{Path, PathBuf},
     process::{Child, ChildStdin, ChildStdout, Command, Stdio},
     thread,
@@ -19,7 +19,6 @@ use std::{
 };
 
 use anyhow::{anyhow, Context, Result};
-use assert_cmd::cargo::cargo_bin;
 use serde_json::{json, Value};
 use tempfile::TempDir;
 
@@ -39,14 +38,8 @@ impl Drop for LspGuard {
 }
 
 fn spawn_lsp(root: &Path) -> Result<LspGuard> {
-    let bin = cargo_bin("deslop-lsp");
-    let mut child = Command::new(bin)
-        .arg(root)
-        .stdin(Stdio::piped())
-        .stdout(Stdio::piped())
-        .stderr(Stdio::null())
-        .spawn()
-        .context("spawn deslop-lsp")?;
+    let mut child =
+        deslop_test_support::spawn_deslop_lsp(root, Stdio::null()).context("spawn deslop-lsp")?;
     let mut stdin = child.stdin.take().context("lsp stdin")?;
     let mut stdout = BufReader::new(child.stdout.take().context("lsp stdout")?);
     lsp_handshake(&mut stdin, &mut stdout).context("lsp handshake")?;
@@ -73,29 +66,11 @@ fn lsp_handshake(stdin: &mut ChildStdin, stdout: &mut BufReader<ChildStdout>) ->
 }
 
 fn write_lsp_frame(stdin: &mut ChildStdin, payload: &str) -> Result<()> {
-    let header = format!("Content-Length: {}\r\n\r\n", payload.len());
-    stdin.write_all(header.as_bytes())?;
-    stdin.write_all(payload.as_bytes())?;
-    stdin.flush()?;
-    Ok(())
+    deslop_test_support::write_lsp_frame(stdin, payload)
 }
 
 fn read_lsp_frame(reader: &mut BufReader<ChildStdout>) -> Result<Value> {
-    let mut content_length: Option<usize> = None;
-    loop {
-        let mut line = String::new();
-        let _bytes = reader.read_line(&mut line)?;
-        if line == "\r\n" {
-            break;
-        }
-        if let Some(rest) = line.strip_prefix("Content-Length: ") {
-            content_length = Some(rest.trim().parse::<usize>()?);
-        }
-    }
-    let length = content_length.ok_or_else(|| anyhow!("missing Content-Length"))?;
-    let mut buf = vec![0_u8; length];
-    reader.read_exact(&mut buf)?;
-    Ok(serde_json::from_slice(&buf)?)
+    deslop_test_support::read_lsp_frame(reader)
 }
 
 fn wait_for_socket(path: &Path) -> Result<()> {
