+++ crates/deslop-test-support/Cargo.toml
@@ -1,0 +1,17 @@
+# agent-pmo:9a71cbf
+[package]
+name = "deslop-test-support"
+version.workspace = true
+edition.workspace = true
+rust-version.workspace = true
+license.workspace = true
+repository.workspace = true
+publish = false
+
+[dependencies]
+anyhow.workspace = true
+assert_cmd.workspace = true
+serde_json.workspace = true
+
+[lints]
+workspace = true
