+++ crates/deslop/Cargo.toml
@@ -22,6 +22,7 @@ tracing-subscriber.workspace = true
 [dev-dependencies]
 assert_cmd.workspace = true
 deslop-core = { path = "../deslop-core", features = ["test-support"] }
+deslop-test-support = { path = "../deslop-test-support" }
 predicates.workspace = true
 tempfile.workspace = true
